Revert "FPS Counter Upgrade" PhotonSettings.asset
This reverts commit 90a2360b2c9a4f27bc142b114dd421f9c9666fc6 for photon settings.

Also move the extensions in extensions classes and fix the code smell relative to em
