--- Assets/Assets.index
@@ -1,19 +1,0 @@
-{
-    "name": "Assets",
-    "type": "asset",
-    "roots": [],
-    "includes": [],
-    "excludes": [
-        "Temp/",
-        "External/"
-    ],
-    "baseScore": 100,
-    "options": {
-        "disabled": false,
-        "files": true,
-        "directories": false,
-        "types": true,
-        "properties": false,
-        "dependencies": false
-    }
-}
