--- Assets/Assets.index.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 2bbd09f9a8fd86443981989a8a8bbfc2
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 11500000, guid: 595da1eac225fa84fb9c2d465cc69e8b, type: 3}
