+++ Assets/Material/DayNightCycle/SkyBoxes/Procedural Skybox.mat
@@ -21,7 +21,7 @@ Material:
     m_TexEnvs: []
     m_Floats:
     - _Angle: -180
-    - _AtmosphereThickness: 1
+    - _AtmosphereThickness: 1.0002785
     - _Exposure: 1.3
     - _Frequency: 100
     - _Gain: 3
