+++ Assets/Photon Unity Networking/Resources/PhotonServerSettings.asset
@@ -15,14 +15,14 @@ MonoBehaviour:
   AppID: 
   VoiceAppID: 
   ChatAppID: 
-  HostType: 0
+  HostType: 2
   PreferredRegion: 0
   EnabledRegions: -1
   Protocol: 0
   ServerAddress: 
   ServerPort: 5055
   VoiceServerPort: 5055
-  JoinLobby: 0
+  JoinLobby: 1
   EnableLobbyStatistics: 0
   PunLogging: 0
   NetworkLogging: 1
@@ -43,6 +43,7 @@ MonoBehaviour:
   - DestroyRpc
   - endMovingRock
   - GetHit
+  - getRacingResult
   - Grab
   - grabbedTargetEscape
   - GrabEscapeRpc
@@ -78,6 +79,7 @@ MonoBehaviour:
   - NetPlayAnimation
   - netPlayAnimationAt
   - NetPlayAnimationAt
+  - netRefreshRacingResult
   - NetSetIsGrabbedFalse
   - netShowDamage
   - netTauntAttack
@@ -96,7 +98,6 @@ MonoBehaviour:
   - PunPickup
   - PunPickupSimple
   - PunRespawn
-  - RacingFinishRpc
   - RacingStartRpc
   - RefreshCaptureScore
   - removeMe
@@ -121,7 +122,6 @@ MonoBehaviour:
   - SetMyPhotonCamera
   - SetMyTeam
   - setPhase
-  - SetRacingResultsRpc
   - setSize
   - SetSize
   - setTeamRPC
@@ -142,4 +142,6 @@ MonoBehaviour:
   - updateKillInfo
   - verifyPlayerHasLeft
   - WhoIsMyErenTitan
+  - RacingFinishRpc
+  - SetRacingResultsRpc
   DisableAutoOpenWizard: 1
