+++ Assets/Resources/Day and Night Controller.prefab
@@ -380,7 +380,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 5156324855798302335}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 122.98607, y: 81.09024, z: 122.98608}
+  m_LocalPosition: {x: 865.55756, y: 358.92972, z: -1507.9188}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 456198}
