+++ Assets/Scripts/Extensions/DictionaryExtensions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fb0e979a5a335c2428c9147b047b0e0c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
