+++ Assets/Scripts/Legacy/RCextensions.cs
@@ -172,27 +172,6 @@ public static class RCextensions
         return 0f;
     }
 
-    public static T SafeGet<T, T1, T2>(this IDictionary<T1, T2> h, T1 key) where T : new()
-    {
-        if (!h.ContainsKey(key))
-            return new T();
-        object o = h[key];
-        return (o != null && o is T ? (T) o : new T());
-    }
-
-    public static T SafeGet<T, T1, T2>(this IDictionary<T1, T2> h, T1 key, T defaultValue)
-    {
-        if (!h.ContainsKey(key))
-            return defaultValue;
-        object o = h[key];
-        return (o != null && o is T ? (T) o : defaultValue);
-    }
-
-    public static bool Between<T>(this T val, T min, T max) where T : IComparable
-    {
-        return (val.CompareTo(max) + min.CompareTo(val))<0;
-    }
-
     public static int returnIntFromObject(object obj)
     {
         if ((obj != null) && (obj is int))
